+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 1.8 (unreleased)
 ----------------
 
-- Nothing changed yet.
-
+- Make REQUEST available thru self.request.
+  [gbastien]
 
 1.7 (2018-06-26)
 ----------------