+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
 - Added methods `removeUnusedColumns` and `removeUnusedIndexes` to easily remove
   columns or indexes from portal_catalog.
   [gbastien]
+- Possibility to run specific upgrade steps
+  [sgeulette]
 
 1.7 (2018-06-26)
 ----------------